pyproject.toml
[project]
name = "flood-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.0.4",
    "fastapi>=0.122.0",
    "geopandas>=1.1.1",
    "ipykernel>=7.1.0",
    "line-bot-sdk>=3.21.0",
    "osmnx>=2.0.7",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
